Include default timeout value from interface
Both the interface (`Interop\Queue\PsrConsumer`) and the implementation (for example `Enqueue\AmqpBunny\AmqpConsumer`) seem to have a default value for the `receive()` method's first and only parameter `$timeout`. This default value (`0`) isn't included in the `@method` annotation in `Interop\Queue\PsrConsumer\AmqpConsumer` which makes code completion in an IDE give a false positive when leaving out the (optional) parameter.

Including this default value in the annotation fixes this problem. There is no effect on the actual functionality or operation of this interop library or any of it's decendants.